// model/customerFiscalEntitiesRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CustomerFiscalEntitiesRequestAddress } from './customerFiscalEntitiesRequestAddress';

export class CustomerFiscalEntitiesRequest {
    'address': CustomerFiscalEntitiesRequestAddress;
    'taxId'?: string;
    'email'?: string;
    'phone'?: string;
    'metadata'?: { [key: string]: object; };
    'companyName'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "address",
            "baseName": "address",
            "type": "CustomerFiscalEntitiesRequestAddress"
        },
        {
            "name": "taxId",
            "baseName": "tax_id",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "phone",
            "baseName": "phone",
            "type": "string"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: object; }"
        },
        {
            "name": "companyName",
            "baseName": "company_name",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CustomerFiscalEntitiesRequest.attributeTypeMap;
    }
}