// model/customerPaymentMethodsRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { PaymentMethodCardRequest } from './paymentMethodCardRequest';
import { PaymentMethodCashRequest } from './paymentMethodCashRequest';
import { PaymentMethodSpeiRequest } from './paymentMethodSpeiRequest';

export class CustomerPaymentMethodsRequest {
    /**
    * Type of payment method
    */
    'type': string;
    /**
    * Token id that will be used to create a \"card\" type payment method. See the (subscriptions)[https://developers.conekta.com/v2.1.0/reference/createsubscription] tutorial for more information on how to tokenize cards.
    */
    'tokenId': string;
    'expiresAt'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "string"
        },
        {
            "name": "tokenId",
            "baseName": "token_id",
            "type": "string"
        },
        {
            "name": "expiresAt",
            "baseName": "expires_at",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CustomerPaymentMethodsRequest.attributeTypeMap;
    }
}