// model/customerShippingContacts.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CustomerShippingContactsAddress } from './customerShippingContactsAddress';

/**
* [Shipping](https://developers.conekta.com/v2.1.0/reference/createcustomershippingcontacts) details, required in case of sending a shipping. If we do not receive a shipping_contact on the order, the default shipping_contact of the customer will be used.
*/
export class CustomerShippingContacts {
    /**
    * Phone contact
    */
    'phone'?: string;
    /**
    * Name of the person who will receive the order
    */
    'receiver'?: string;
    /**
    * The street names between which the order will be delivered.
    */
    'betweenStreets'?: string;
    'address': CustomerShippingContactsAddress;
    'parentId'?: string;
    '_default'?: boolean | null;
    'deleted'?: boolean | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "phone",
            "baseName": "phone",
            "type": "string"
        },
        {
            "name": "receiver",
            "baseName": "receiver",
            "type": "string"
        },
        {
            "name": "betweenStreets",
            "baseName": "between_streets",
            "type": "string"
        },
        {
            "name": "address",
            "baseName": "address",
            "type": "CustomerShippingContactsAddress"
        },
        {
            "name": "parentId",
            "baseName": "parent_id",
            "type": "string"
        },
        {
            "name": "_default",
            "baseName": "default",
            "type": "boolean"
        },
        {
            "name": "deleted",
            "baseName": "deleted",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return CustomerShippingContacts.attributeTypeMap;
    }
}