// model/subscriptionRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* It is a parameter that allows to identify in the response, the detailed content of the plans to which the client has subscribed
*/
export class SubscriptionRequest {
    'planId': string;
    'cardId'?: string;
    'trialEnd'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "planId",
            "baseName": "plan_id",
            "type": "string"
        },
        {
            "name": "cardId",
            "baseName": "card_id",
            "type": "string"
        },
        {
            "name": "trialEnd",
            "baseName": "trial_end",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return SubscriptionRequest.attributeTypeMap;
    }
}